% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAbstracts.R
\name{Get.Abstracts}
\alias{Get.Abstracts}
\title{Get Abstracts of user-defined search query}
\usage{
Get.Abstracts(Searchquery, Output = NULL)
}
\arguments{
\item{Searchquery}{input a text query for example "Bioinformatics", "Bioinformatics or machine learning", "Cancer and drug targets"}

\item{Output}{path to save file containig results returened by the function.}
}
\description{
Get Abstracts of user-defined search query
}
\examples{
Abstracts <- Get.Abstracts(Searchquery = "physics and deep learning and proteomics")

}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
