% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPublications.R
\name{Plot.papertype}
\alias{Plot.papertype}
\title{Plot papers type returned by "Get.Publication.info" Function}
\usage{
Plot.papertype(Publications, directorypath  = NULL)
}
\arguments{
\item{Publications}{Dataframe returned by "Get.Publication.info" Function.}

\item{directorypath}{path to save plot generated by the function.}
}
\description{
Plot papers type returned by "Get.Publication.info" Function
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
