% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPublications.R
\name{Plot.TopPapers}
\alias{Plot.TopPapers}
\title{Plot Top cited publications from data returned by "Get.Publication.info" Function}
\usage{
Plot.TopPapers(Publications, Top = 10, directorypath  = NULL)
}
\arguments{
\item{Publications}{Dataframe returned by "Get.Publication.info" Function.}

\item{Top}{Number of publications to be visulaized sorted by citations.}

\item{directorypath}{path to save plot generated by the function.}
}
\description{
Plot Top cited publications from data returned by "Get.Publication.info" Function
}
\author{
Mohmed Soudy \email{Mohamed.soudy@57357.com} and Ali Mostafa \email{ali.mo.anwar@std.agr.cu.edu.eg}
}
